Implement Set with an UnbalancedTreeSet wrapper
The wrapper behaves like a fully-fledged Set, but underneath the hood it
just wraps a Tree, using the operations defined on the UnbalancedTreeSet
object.

I think this does a good job of being idiomatic to Scala, while also
retaining the semantics and "feel" of the PFDS implementation.

Also remove the TreeSet trait from Tree.scala, making it just a datatype
set of a classes.